requirements.txt
discord.py
osuapi